Refactor modules and remove legacy
