+++ app/core/indicators.py
