+++ app/core/indicators_vectorized.py
