+++ app/strategy/manager.py
