+++ app/symbol_engine_manager.py
@@ -1,14 +1,14 @@
 import asyncio
+import time
 from types import SimpleNamespace as NS
 
 
 from app.symbol_engine import SymbolEngine
 from app.hybrid_strategy_engine import HybridStrategyEngine
 from app.config import settings
-from app.command_listener import telegram_command_listener
-from app.exchange import BybitClient
+from app.ws_client import WsClient
 from app.notifier import notify_telegram
-from app.risk_guard import RiskGuard  # new class
+from app.risk_guard import RiskGuard
 
 class SymbolEngineManager:
     def __init__(self, symbols: list[str]):
@@ -16,11 +16,11 @@ class SymbolEngineManager:
         self.tasks: dict[str, asyncio.Task] = {}
         self.engines: dict[str, SymbolEngine] = {}
         self.account = NS(equity_usd=0.0, open_positions=[])
-        self.guard = RiskGuard(self.account)
+        self.risk_guard = RiskGuard(self.account)
         if settings.risk.max_open_positions:
-            self.guard.MAX_POSITIONS = settings.risk.max_open_positions
+            self.risk_guard.MAX_POSITIONS = settings.risk.max_open_positions
         if settings.trading.max_position_risk_percent:
-            self.guard.TOTAL_RISK_CAP_PCT = settings.trading.max_position_risk_percent
+            self.risk_guard.TOTAL_RISK_CAP_PCT = settings.trading.max_position_risk_percent
 
     async def _run_engine(self, symbol: str, ref_symbol: str | None = None):
         engine_cls = (
@@ -72,30 +72,28 @@ class SymbolEngineManager:
 
         # shared WS connections ----------------------------------------
         self.tasks["orderbook"] = asyncio.create_task(
-            BybitClient.ws_multi(active, "orderbook.50", self._on_orderbook)
+            WsClient.ws_multi(active, "orderbook.50", self._on_orderbook)
         )
         self.tasks["trades"] = asyncio.create_task(
-            BybitClient.ws_multi(active, "publicTrade", self._on_trades)
+            WsClient.ws_multi(active, "publicTrade", self._on_trades)
         )
 
-        self.tasks["cmd"] = asyncio.create_task(telegram_command_listener())
         await asyncio.gather(*self.tasks.values())
 
     async def _maybe_open_position(self, engine: SymbolEngine, direction: str, price: float) -> bool:
-        risk_pct = settings.trading.initial_risk_percent
-        guard = self.guard
-        if settings.risk.enable_daily_trades_guard:
-            if guard.today_trades >= settings.risk.daily_trades_limit:
-                print(f"🚫 Daily trades limit ({guard.today_trades})")
-                return False
-        if not guard.allow_new_position(risk_pct):
-            print("🚫 Portfolio risk cap hit")
+        if not self.risk_guard.allow_new_position(engine.symbol, 0, direction):
+            print(f"[{engine.symbol}] 🚫 Blocked by RiskGuard")
             return False
         if engine.entry_order_id is not None or engine.risk.position.qty > 0:
             return False
         await engine._open_position(direction, price)
-        self.account.open_positions.append(NS(symbol=engine.symbol, risk_pct=risk_pct))
-        guard.inc_trade()
+        self.risk_guard.register_trade(
+            engine.symbol,
+            engine.risk.position.qty,
+            engine.risk.position.side or direction,
+            time.time(),
+            engine.risk.position.qty * price,
+        )
         return True
 
     def position_closed(self, engine: SymbolEngine) -> None:
