+++ tests/test_ohlc.py
@@ -1,9 +1,13 @@
-import asyncio
-from legacy.symbol_engine import SymbolEngine
-from tests.conftest import feed_trades
+from app.core.data import OHLCCollector
+
+
+async def feed_collector(col, trades):
+    for t in trades:
+        col.on_trade(t.price, t.qty, t.ts)
 
 
 def test_one_bar(fixture_trades_5m):
-    se = SymbolEngine("BTCUSDT")
-    asyncio.run(feed_trades(se, fixture_trades_5m))
-    assert se.ohlc.last_bar.close == fixture_trades_5m[-1].price
+    col = OHLCCollector()
+    import asyncio
+    asyncio.run(feed_collector(col, fixture_trades_5m))
+    assert col.last_bar.close == fixture_trades_5m[-1].price
